Update protos for splits and joins.
